move internal signer to bt pkg
