+++ bt/InternalSigner.go
@@ -1,11 +1,10 @@
-package sig
+package bt
 
 import (
 	"encoding/hex"
 
 	"github.com/bitcoinsv/bsvd/bsvec"
 
-	"github.com/libsv/libsv/bt"
 	"github.com/libsv/libsv/bt/sighash"
 	"github.com/libsv/libsv/crypto"
 	"github.com/libsv/libsv/script"
@@ -21,7 +20,7 @@ type InternalSigner struct {
 
 // Sign a transaction at a given input index using the PrivateKey passed in through the
 // InternalSigner struct.
-func (is *InternalSigner) Sign(index uint32, unsignedTx *bt.Tx) (*bt.Tx, error) {
+func (is *InternalSigner) Sign(index uint32, unsignedTx *Tx) (*Tx, error) {
 	if is.SigHashFlag == 0 {
 		is.SigHashFlag = sighash.AllForkID
 	}
@@ -49,7 +48,7 @@ func (is *InternalSigner) Sign(index uint32, unsignedTx *bt.Tx) (*bt.Tx, error) 
 // SignAuto goes through each input of the transaction and automatically
 // signs the P2PKH inputs that it is able to sign using the specific
 // PrivateKey passed in through the InternalSigner struct.
-func (is *InternalSigner) SignAuto(unsignedTx *bt.Tx) (*bt.Tx, error) {
+func (is *InternalSigner) SignAuto(unsignedTx *Tx) (*Tx, error) {
 	if is.SigHashFlag == 0 {
 		is.SigHashFlag = sighash.AllForkID
 	}
